refactor(dropdown): Be more similar to angular-ui-bootstrap
As #3 brings up, the design of dropdownToggle makes it difficult to use
within ngRepeat directives. This patch takes code from angular-ui-bootstrap
and replaces the dropdownToggle directive with dropdown, which is similar
(but not identical) to the one in angular-ui-bootstrap. I tried to make
it as similar as possible.

The new dropdown directive is incompatible with dropdownToggle, so I
removed dropdownToggle.