Allow dropdownToggle directive to inherit parent scope
When paired with another directive on the same element, Angular may complain about multiple directives trying to create a new scope.  This directive (dropdownToggle) doesn't seem to really require its own scope, so we should be able to remove it and get the information we need from the "attrs" parameter instead.